templates
- with replaceable names supporting dotnet new --name parameter
- website template structure
- website template compilable and with replaceable namespaces
- lib -> class library
- lib also builds a nuget package
