+++ src/Content/peachpie-classlibrary/.template.config/template.json
@@ -1,0 +1,13 @@
+{
+  "author": "Peachpie",
+  "classifications": [ "Library" ],
+  "name": "Peachpie Class library",
+  "identity": "Peachpie.Templates.ClassLibrary",
+  "shortName": "peachpie-classlibrary",
+  "sourceName": "PhpLib1",
+  "preferNameDirectory" : "true",
+  "groupIdentity": "Peachpie.Templates.ClassLibrary", // .NET CLI 1.x workaround (https://github.com/dotnet/templating/issues/649)
+  "tags": {
+    "language": "PHP"
+  }
+}
