+++ src/Content/peachpie-classlibrary/lib.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Library {
+
+	/** A greeter class. */
+	class Greeter {
+		/** A method that outputs hello. */
+		public function Greet(){
+			echo "Hello world!";
+		}
+	}
+}
