+++ src/Content/peachpie-console/.template.config/template.json
@@ -4,7 +4,8 @@
   "name": "Peachpie console application",
   "identity": "Peachpie.Templates.ConsoleApp",
   "shortName": "peachpie-console",  
-  "sourceName": "MyTemplate", // .NET CLI 1.x workaround (https://github.com/dotnet/templating/issues/650)
+  "sourceName": "PhpConsole1",
+  "preferNameDirectory" : "true",
   "groupIdentity": "Peachpie.Templates.ConsoleApp", // .NET CLI 1.x workaround (https://github.com/dotnet/templating/issues/649)
   "tags": {
     "language": "PHP"
