+++ src/Content/peachpie-console/PhpConsole1.msbuildproj
@@ -3,7 +3,8 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp1.0</TargetFramework>
-    <Description>Simpliest PHP project for .NET Core, a console app.</Description>
+	<StartupObject>program.php</StartupObject>
+    <Description>Simpliest PHP project for .NET Core, a console application.</Description>
   </PropertyGroup>
 
   <ItemGroup>
