--- src/Content/peachpie-phplib/.template.config/template.json
@@ -1,12 +1,0 @@
-{
-  "author": "Peachpie",
-  "classifications": [ "Library" ],
-  "name": "Peachpie library",
-  "identity": "Peachpie.Templates.PhpLib",
-  "shortName": "peachpie-phplib",
-  "sourceName": "MyTemplate", // .NET CLI 1.x workaround (https://github.com/dotnet/templating/issues/650)
-  "groupIdentity": "Peachpie.Templates.PhpLib", // .NET CLI 1.x workaround (https://github.com/dotnet/templating/issues/649)
-  "tags": {
-    "language": "PHP"
-  }
-}
