--- src/Content/peachpie-phplib/lib.php
@@ -1,7 +1,0 @@
-<?php
-
-class Greeter {
-    public function Greet(){
-        echo "Hello world!";
-    }
-}
