+++ src/Content/peachpie-web/MyWebsite1.Server/MyWebsite1.Server.csproj
@@ -10,11 +10,11 @@
     <PackageReference Include="Microsoft.AspNetCore.Session" Version="1.1.1" />
     <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="1.1.1" />
     <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="1.1.1" />
-    <PackageReference Include="Peachpie.NETCore.Web" Version="0.6.1-*" />
+    <PackageReference Include="Peachpie.NETCore.Web" Version="0.7.0-*" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include=".\website\website.msbuildproj" />
+    <ProjectReference Include="..\MyWebsite1\MyWebsite1.msbuildproj" />
   </ItemGroup>
 
 </Project>
