+++ src/Content/peachpie-web/MyWebsite1.sln
@@ -1,0 +1,48 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26430.16
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{13B669BE-BB05-4DDF-9536-439F39A36129}") = "MyWebsite1", "MyWebsite1\MyWebsite1.msbuildproj", "{419BB192-DEA4-498A-8FFD-093F39CC884B}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyWebsite1.Server", "MyWebsite1.Server\MyWebsite1.Server.csproj", "{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{419BB192-DEA4-498A-8FFD-093F39CC884B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{419BB192-DEA4-498A-8FFD-093F39CC884B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{419BB192-DEA4-498A-8FFD-093F39CC884B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{419BB192-DEA4-498A-8FFD-093F39CC884B}.Debug|x64.Build.0 = Debug|Any CPU
+		{419BB192-DEA4-498A-8FFD-093F39CC884B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{419BB192-DEA4-498A-8FFD-093F39CC884B}.Debug|x86.Build.0 = Debug|Any CPU
+		{419BB192-DEA4-498A-8FFD-093F39CC884B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{419BB192-DEA4-498A-8FFD-093F39CC884B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{419BB192-DEA4-498A-8FFD-093F39CC884B}.Release|x64.ActiveCfg = Release|Any CPU
+		{419BB192-DEA4-498A-8FFD-093F39CC884B}.Release|x64.Build.0 = Release|Any CPU
+		{419BB192-DEA4-498A-8FFD-093F39CC884B}.Release|x86.ActiveCfg = Release|Any CPU
+		{419BB192-DEA4-498A-8FFD-093F39CC884B}.Release|x86.Build.0 = Release|Any CPU
+		{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}.Debug|x64.Build.0 = Debug|Any CPU
+		{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}.Debug|x86.Build.0 = Debug|Any CPU
+		{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}.Release|x64.ActiveCfg = Release|Any CPU
+		{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}.Release|x64.Build.0 = Release|Any CPU
+		{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}.Release|x86.ActiveCfg = Release|Any CPU
+		{E05A30A4-4806-4E9F-9404-1DBD6B3571A5}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
