+++ src/Content/peachpie-web/MyWebsite1/MyWebsite1.msbuildproj
@@ -2,8 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Library</OutputType>
-    <TargetFrameworks>netstandard1.5;net46</TargetFrameworks>
-    <Description>.NET library in PHP</Description>
+    <TargetFramework>netstandard1.6</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
