--- src/Content/peachpie-web/Program.cs
@@ -1,23 +1,0 @@
-using Microsoft.AspNetCore.Hosting;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace peachserver
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            var host = new WebHostBuilder()
-                .UseKestrel()
-                .UseUrls("http://*:5004/")
-                .UseStartup<Startup>()
-                .Build();
-
-            host.Run();
-        }
-    }
-}
